make tokenizer typesafe and move to separate file
