+++ packages/core/src/helper/index.ts
@@ -2,4 +2,4 @@ export * from './detectMobileBrowser';
 export * from './urlRegex';
 export * from './emailRegex';
 export * from './floating-menu';
-
+export * from './tokenizer';
